requirement.txt
# Core
opencv-python>=4.5.5,<4.9.0
numpy==1.26.4

# AI Models
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
ultralytics==8.0.196
insightface==0.7.3
efficientnet-pytorch==0.7.1
onnxruntime-gpu==1.16.3  # Fixed typo (changed =- to ==)

# Web
flask==2.3.2
flask-socketio==5.3.6
pygame==2.5.2  # For alarms